reorganized tests (again)
